handling mount errors

+++ download.sh
@@ -12,14 +12,25 @@ if [ ! -f "$home/$tarballs/$filename" ]; then
 	wget $1
 	
 	if [ ! -f "$home/$filename" ]; then
-		echo "ERR, couldn't find or download $filename"
+		echo "ERROR, couldn't find or download $filename"
 		exit 1
 	fi
 
 	mv $filename $tarballs
 fi
 
+if [ -n "`mount | grep /dev/sda1`" ]; then
+	echo "WARNING, device already mounted"
+	exit 0
+fi
+
 sudo mount /dev/sda1 $usb
+
+if [ $? -ne 0 ]; then
+	echo "ERROR, mount failed. Is the USB plugged in?"
+	exit 1
+fi
+
 sudo rm -rf $usb/*.tar.xz*
 echo "Copying tarball to USB"
 sudo cp $tarballs/$filename $usb
